requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai-whisper
pydub==0.25.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
aiofiles==23.2.1